"14-07-24"

+++ journalApp/journalApp/src/main/java/com/journalapp/journalApp/services/JournalEntryService.java
@@ -6,12 +6,15 @@ import com.journalapp.journalApp.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
 @Component
+@Service
 public class JournalEntryService {
     @Autowired
     private JournalEntryRepository journalEntryRepository;
@@ -19,14 +22,18 @@ public class JournalEntryService {
     @Autowired
     private UserService userService;
 
+    @Transactional
     public void saveEntry(JournalEntry journalEntry, String userName){
-
-          User user=   userService.findByUserName(userName);
-          journalEntry.setDate(LocalDateTime.now());
-         JournalEntry saved= journalEntryRepository.save(journalEntry);
-         user.getJournalEntries().add(saved);
-         userService.saveEntry(user);
-
+        try {
+            User user = userService.findByUserName(userName);
+            journalEntry.setDate(LocalDateTime.now());
+            JournalEntry saved = journalEntryRepository.save(journalEntry);
+            user.getJournalEntries().add(saved);
+            userService.saveNewUser(user);
+        }catch(Exception e){
+            System.out.println(e);
+            throw new RuntimeException("an error occurred while saving the entry",e);
+        }
     }
     public void saveEntry(JournalEntry journalEntry){
 
@@ -42,10 +49,21 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
-        User user=   userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed=false;
+        try {
+            User user = userService.findByUserName(userName);
+             removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveNewUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+
+        } catch(Exception e){
+            System.out.println("Error"+ e);
+            throw new RuntimeException("An error occurred while deleting the entry",e);
+        }
+        return removed;
     }
 }
